Revert "Squashed commit of the following:"
This reverts commit 41e24df4b5cfb071d1a47c60123def00307fb46d.

+++ src/navigation/orbitalnavigator_lua.inl
@@ -36,7 +36,7 @@ namespace {
 
     double is = node->interactionSphere();
     global::navigationHandler->orbitalNavigator().setMinimumAllowedDistance(
-        static_cast<float>(is * multiplier)
+        is * multiplier
     );
 }
 
@@ -52,7 +52,7 @@ namespace {
 
     double is = node->interactionSphere();
     global::navigationHandler->orbitalNavigator().setMaximumAllowedDistance(
-        static_cast<float>(is * multiplier)
+        is * multiplier
     );
 }
